# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Laravel",
        "version": "0.1"
    },
    "paths": {
        "/usuarios": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Obter todos os usuários",
                "operationId": "90a70f9fe889fa61b696cc4e05f064aa",
                "responses": {
                    "200": {
                        "description": "Lista de usuários"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/usuarios/id": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Obter um usuário pelo ID",
                "operationId": "207b5772615e81986bd889844e30ae8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do usuário"
                    },
                    "401": {
                        "description": "No query results for model [App\\\\Models\\\\User] id"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Autenticar usuário",
                "operationId": "0fc9610184083ac6dddab2c63139225a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token"
                    },
                    "401": {
                        "description": "As credenciais fornecidas estão incorretas."
                    }
                }
            }
        }
    }
}